Move and reduce event tests (#35693)
